Raise when path is missing
When trying to add a route handler in bottle without specifying a path, bottle
tries to do something clever by introspecting the function and generating a
path that matches its name and argspec. However, with class-based route, the
class object is *not* the actual callback, so this mechanism fails. Since we
can't get as smart as bottle (there could be any number of methods that will
actually handle the request), there does not seem to be a viable workaround.

This patch causes the route() method to raise when path is unspecified so at
least we don't have to scratch our heads trying to figure out why the call
failed.

Signed-off-by: Branko Vukelic <branko@outernet.is>